fix: cyclic reference
